Finish paginationApi
