+++ src/js/search.js
@@ -1,14 +1,11 @@
 import getMoviesApi from './getMoviesApi';
 import { appendMoviesMarkup } from './moviesMarkup';
-import { clearGallery } from './moviesMarkup';
-import { clearPagContainer } from './pagination';
-import { createPagination } from './pagination';
+import Pagination from './paginationApi';
 import { Notify } from 'notiflix/build/notiflix-notify-aio';
-import { loadFromTop } from './startPageGalleryRender';
 
 const searchMoviesApi = new getMoviesApi();
+const searchPagination = new Pagination();
 const searchForm = document.querySelector('#search-form');
-const pagContainer = document.querySelector('.pag-container');
 
 searchForm.addEventListener('submit', onSearchFormSubmit);
 
@@ -24,11 +21,17 @@ export function onSearchFormSubmit(evt) {
     );
   }
 
+  function search() {
+    searchMoviesApi.getSearchMovies().then(data => {
+      const movies = data.results;
+
+      appendMoviesMarkup(movies);
+    });
+  }
+
   searchMoviesApi.resetPage();
   searchMoviesApi.getSearchMovies().then(data => {
     const movies = data.results;
-
-    clearGallery();
 
     if (movies.length === 0) {
       return Notify.failure(
@@ -37,8 +40,10 @@ export function onSearchFormSubmit(evt) {
     }
 
     appendMoviesMarkup(movies);
-    clearPagContainer();
-    createPagination(data.total_pages, 5);
+
+    searchPagination.clearPagContainer();
+    searchPagination.create(data.total_pages, 5, searchMoviesApi, search);
+
     if (movies.length === 1) {
       return Notify.success('😮 We have found only one film for your request.');
     }
@@ -47,45 +52,4 @@ export function onSearchFormSubmit(evt) {
       `🎉 Hooray! We have found "${data.total_results}" films with "${query}".🥳`
     );
   });
-
-  pagContainer.addEventListener('click', paginate);
-
-  function paginate(evt) {
-    if (parseInt(evt.target.id) === searchMoviesApi.page) return;
-    if (evt.target.classList.contains('pag-btn')) {
-      searchMoviesApi.page = parseInt(evt.target.id);
-      clearGallery();
-      loadFromTop();
-      searchMoviesApi.getSearchMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-    if (
-      evt.target.classList.contains('move-left') ||
-      evt.target.classList.contains('move-right')
-    ) {
-      searchMoviesApi.page = parseInt(evt.target.id);
-      clearGallery();
-      loadFromTop();
-      searchMoviesApi.getSearchMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-    if (evt.target.classList.contains('to-start')) {
-      searchMoviesApi.page = 1;
-      clearGallery();
-      loadFromTop();
-      searchMoviesApi.getSearchMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-    if (evt.target.classList.contains('to-end')) {
-      searchMoviesApi.page = parseInt(evt.target.id);
-      clearGallery();
-      loadFromTop();
-      searchMoviesApi.getSearchMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-  }
 }
