+++ src/js/startPageGalleryRender.js
@@ -1,60 +1,23 @@
 import getMoviesApi from './getMoviesApi';
 import { appendMoviesMarkup } from './moviesMarkup';
-import { clearGallery } from './moviesMarkup';
-import { clearPagContainer } from './pagination';
-import { createPagination } from './pagination';
+import Pagination from './paginationApi';
 
 const getMovies = new getMoviesApi();
-const pagContainer = document.querySelector('.pag-container');
+const startPagination = new Pagination();
 
 function onPageStart() {
+  function get() {
+    getMovies.getStartMovies().then(movies => {
+      appendMoviesMarkup(movies.results);
+    });
+  }
   getMovies.resetPage();
   getMovies.getStartMovies().then(movies => {
     appendMoviesMarkup(movies.results);
-    clearPagContainer();
-    createPagination(movies.total_pages, 5);
+
+    startPagination.clearPagContainer();
+    startPagination.create(movies.total_pages, 5, getMovies, get);
   });
-
-  pagContainer.addEventListener('click', paginate);
-
-  function paginate(evt) {
-    if (parseInt(evt.target.id) === getMovies.page) return;
-    if (evt.target.classList.contains('pag-btn')) {
-      getMovies.page = parseInt(evt.target.id);
-      clearGallery();
-      loadFromTop();
-      getMovies.getStartMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-    if (
-      evt.target.classList.contains('move-left') ||
-      evt.target.classList.contains('move-right')
-    ) {
-      getMovies.page = parseInt(evt.target.id);
-      clearGallery();
-      loadFromTop();
-      getMovies.getStartMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-    if (evt.target.classList.contains('to-start')) {
-      getMovies.page = 1;
-      clearGallery();
-      loadFromTop();
-      getMovies.getStartMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-    if (evt.target.classList.contains('to-end')) {
-      getMovies.page = parseInt(evt.target.id);
-      clearGallery();
-      loadFromTop();
-      getMovies.getStartMovies().then(movies => {
-        appendMoviesMarkup(movies.results);
-      });
-    }
-  }
 }
 
 function loadFromTop() {
